Resolve PR comments; make isTokenAllowed only a Milestone hook, fix issue with item conversion check
